smart_notes/requirements.txt
pinecone
sentence-transformers
rich
python-dotenv